assets/06_민원처리서버/src/worker/requirements.txt
# =============================================================================
# 민원 처리 서버 (자산 06) Python 패키지
# =============================================================================
# [취약점] VULN-06-02: Pillow 8.4.0 — CVE-2022-22815/22816/22817
# 올바른 구현: Pillow>=10.0.0
# [취약점] VULN-06-01 관련: LibreOffice 6.4.7은 시스템 패키지로 설치 (setup.sh 참조)

celery==5.3.6
redis==5.0.0
boto3==1.34.0
psycopg2-binary==2.9.9
python-magic==0.4.27
Pillow==8.4.0              # [취약점] VULN-06-02: CVE-2022-22815/16/17 (안전 버전: >=10.0.0)
python-dotenv==1.0.0
structlog==24.1.0
